moved algorithm for generating signature out into a new project called apysigner.
Did this so it can be re-used for other API signing stuff independent of Django.

Fixed a bug where the old signer was re-arranging the order of the GET parameters which would mess up the signature if the signature creator didnt make that same mistake. Because of this, it might be a backwards incompatable change.